chore(actionpack): wave 2 — action_dispatch mechanical moves to Rails-mirrored paths
Move actiondispatch source files to their canonical Rails-mirrored
subdirectories under http/, middleware/, middleware/session/, and
request/. Re-export stubs at old paths preserve backward compat for
dispatch/ tests; full test-file relocation and dispatch/ cleanup is
deferred to Wave 2b.

Moved:
- request.ts, response.ts, mime-type.ts, uploaded-file.ts,
  dispatch/header.ts, content-security-policy.ts,
  permissions-policy.ts → http/
- cookies.ts, flash.ts, exception-wrapper.ts → middleware/
- session/cookie-store.ts → middleware/session/
- dispatch/request/session.ts → request/

Follow-up (Wave 2b): move dispatch/*.test.ts to correct subdirs,
delete empty dispatch/ dir, dedup request-forgery-protection and
http-authentication between actiondispatch root and actioncontroller/metal.
