+++ packages/actionpack/src/actioncontroller/test-case.ts
@@ -33,10 +33,10 @@
  *   });
  */
 
-import { Request } from "../actiondispatch/request.js";
-import { Response } from "../actiondispatch/response.js";
+import { Request } from "../actiondispatch/http/request.js";
+import { Response } from "../actiondispatch/http/response.js";
 import { Parameters } from "./metal/strong-parameters.js";
-import { FlashHash } from "../actiondispatch/flash.js";
+import { FlashHash } from "../actiondispatch/middleware/flash.js";
 import type { Metal } from "./metal.js";
 
 type ControllerClass = new () => Metal;
