+++ packages/actionpack/src/actiondispatch/middleware/session/index.ts
@@ -1,1 +1,6 @@
-export {};
+export {
+  CookieStore,
+  CookieOverflow,
+  type CookieStoreOptions,
+  type SessionData,
+} from "./cookie-store.js";
