+++ packages/actionpack/src/actiondispatch/middleware/show-exceptions.ts
@@ -9,7 +9,7 @@
 
 import type { RackApp, RackEnv, RackResponse } from "@blazetrails/rack";
 import { bodyFromString } from "@blazetrails/rack";
-import { ExceptionWrapper } from "../exception-wrapper.js";
+import { ExceptionWrapper } from "./exception-wrapper.js";
 
 export type ShowExceptionsMode = "all" | "rescuable" | "none";
