+++ packages/actionpack/src/actiondispatch/request/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { Session, type SessionStore } from "./session.js";
